refactor: Extract Command building logic into separate components (#1909)
Extract placeholder handling and command parts logic from Command.cs
into dedicated single-responsibility services:

- IPlaceholderHandler/PlaceholderHandler: Handles placeholder replacement
  in command templates (e.g., <argument> or [<optional>] syntax)
- ICommandPartsProvider/CommandPartsProvider: Extracts raw command parts
  from CliCommand, CliSubCommand, and CliCommandAlias attributes

Command.cs now acts as an orchestrator that coordinates:
- ICommandPartsProvider: Get raw command parts (subcommands)
- IPlaceholderHandler: Replace placeholders with actual values
- ICommandModelProvider: Get command model for property arguments
- ICommandArgumentBuilder: Build arguments from command model

This improves testability by allowing individual components to be unit
tested in isolation, and follows the Single Responsibility Principle.

Fixes #1909
